Update abbreviations to only load if not set
Add in fish function, `abbr_set`, that checks `fish_abbr_set` before
refreshing all abbreviations

+++ fish/.config/fish/conf.d/abbr.fish
@@ -1,46 +1,48 @@
-# shell
+if not set -q fish_abbr_set
+  # shell
 
-# git
-abbr --add g    'git'
-abbr --add ga   'git add'
-abbr --add gap  'git add -p'
+  # git
+  abbr --add g    'git'
+  abbr --add ga   'git add'
+  abbr --add gap  'git add -p'
 
-abbr --add gb   'git branch'
+  abbr --add gb   'git branch'
 
-abbr --add gc   'git commit --verbose'
-abbr --add gca  'git commit -a --verbose'
-abbr --add gcam 'git commit -am'
-abbr --add gcm  'git commit -m'
-abbr --add gm   'git commit --amend --verbose'
+  abbr --add gc   'git commit --verbose'
+  abbr --add gca  'git commit -a --verbose'
+  abbr --add gcam 'git commit -am'
+  abbr --add gcm  'git commit -m'
+  abbr --add gm   'git commit --amend --verbose'
 
-abbr --add gco  'git checkout'
+  abbr --add gco  'git checkout'
 
-abbr --add gd   'git diff'
-abbr --add gdc  'git diff --cached'
-abbr --add gdd  'git diff --color-words'
-abbr --add gddc 'git diff --color-words --cached'
-abbr --add gds  'git diff --stat'
-abbr --add gdsc 'git diff --stat --cached'
+  abbr --add gd   'git diff'
+  abbr --add gdc  'git diff --cached'
+  abbr --add gdd  'git diff --color-words'
+  abbr --add gddc 'git diff --color-words --cached'
+  abbr --add gds  'git diff --stat'
+  abbr --add gdsc 'git diff --stat --cached'
 
-abbr --add glg  'git log --graph --max-count=10'
-abbr --add gls  'git log --stat --max-count=10'
-abbr --add gla  'git log --graph --all'
+  abbr --add glg  'git log --graph --max-count=10'
+  abbr --add gls  'git log --stat --max-count=10'
+  abbr --add gla  'git log --graph --all'
 
-abbr --add gp   'git pull'
-abbr --add gpp  'git pull; and git push --force-with-lease'
-abbr --add gpu  'git push --force-with-lease'
+  abbr --add gp   'git pull'
+  abbr --add gpp  'git pull; and git push --force-with-lease'
+  abbr --add gpu  'git push --force-with-lease'
 
-abbr --add grbi 'git rebase -i'
-abbr --add grbc 'git rebase --continue'
-abbr --add grba 'git rebase --abort'
+  abbr --add grbi 'git rebase -i'
+  abbr --add grbc 'git rebase --continue'
+  abbr --add grba 'git rebase --abort'
 
-abbr --add gsta 'git stash'
-abbr --add gstv 'git stash save'
-abbr --add gsts 'git stash show --text'
-abbr --add gstp 'git stash pop'
-abbr --add gstd 'git stash drop'
+  abbr --add gsta 'git stash'
+  abbr --add gstv 'git stash save'
+  abbr --add gsts 'git stash show --text'
+  abbr --add gstp 'git stash pop'
+  abbr --add gstd 'git stash drop'
 
-abbr --add gs   'git status --short'
-abbr --add gsw  'git switch'
-abbr --add gsw- 'git switch -'
-abbr --add gswc 'git switch -c'
+  abbr --add gs   'git status --short'
+  abbr --add gsw  'git switch'
+  abbr --add gsw- 'git switch -'
+  abbr --add gswc 'git switch -c'
+end
